Add float support for courses Fix update courses

+++ api/src/Entity/Course.php
@@ -80,7 +80,7 @@ class Course
     private $image = null;
 
     #[ORM\Column]
-    private ?int $price = null;
+    private ?float $price = null;
 
     #[ORM\Column(length: 255)]
     private ?string $stripeProductId = null;
@@ -291,12 +291,12 @@ class Course
         return $this;
     }
 
-    public function getPrice(): ?int
+    public function getPrice(): ?float
     {
         return $this->price;
     }
 
-    public function setPrice(int $price): self
+    public function setPrice(float $price): self
     {
         $this->price = $price;
